wip: move the type one level up
