+++ src/config/logger.ts
@@ -1,5 +1,5 @@
 import { assertOneOf } from '@/kairos/shared/kernel/assert';
-import { type LogLevel } from '@/kairos/shared/logger/logger.type';
+import { type LogLevel } from '@/kairos/shared/logger.type';
 
 const logLevel = process.env.LOG_LEVEL ?? 'info';
 assertOneOf<LogLevel>(logLevel, 'fatal', 'error', 'warn', 'info', 'debug', 'trace', 'silent');
